+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_files_artifact_expansion_volume.go
@@ -5,19 +5,19 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/file_artifact"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/file_artifact_expansion_volume"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion_volume"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/stacktrace"
 )
 
-func (backend *DockerKurtosisBackend) CreateFileArtifactExpansionVolume(
+func (backend *DockerKurtosisBackend) CreateFilesArtifactExpansionVolume(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
 	serviceGuid service.ServiceGUID,
-	fileArtifactId file_artifact.FilterArtifactID,
+	filesArtifactId files_artifact.FilesArtifactID,
 )(
-	*file_artifact_expansion_volume.FileArtifactExpansionVolume,
+	*files_artifact_expansion_volume.FilesArtifactExpansionVolume,
 	error,
 ) {
 
@@ -26,24 +26,24 @@ func (backend *DockerKurtosisBackend) CreateFileArtifactExpansionVolume(
 		return nil, stacktrace.Propagate(err, "Couldn't get an object attribute provider for enclave '%v'", enclaveId)
 	}
 
-	containerAttrs, err := enclaveObjAttrsProvider.ForFilesArtifactExpansionVolume(serviceGuid, fileArtifactId)
+	volumeAttrs, err := enclaveObjAttrsProvider.ForFilesArtifactExpansionVolume(serviceGuid, filesArtifactId)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred while trying to get the file artifacts expansion volume attributes for user service with GUID '%v'", serviceGuid)
+		return nil, stacktrace.Propagate(err, "An error occurred while trying to get the files artifact expansion volume attributes for user service with GUID '%v' and files artifact ID '%v'", serviceGuid, filesArtifactId)
 	}
-	volumeName := containerAttrs.GetName().GetString()
+	volumeName := volumeAttrs.GetName().GetString()
 	volumeLabels := map[string]string{}
-	for labelKey, labelValue := range containerAttrs.GetLabels() {
+	for labelKey, labelValue := range volumeAttrs.GetLabels() {
 		volumeLabels[labelKey.GetString()] = labelValue.GetString()
 	}
 
-	foundedVolumes, err := backend.dockerManager.GetVolumesByName(ctx, volumeName)
+	foundVolumes, err := backend.dockerManager.GetVolumesByName(ctx, volumeName)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting Docker volumes by name '%v'", volumeName)
 	}
-	if len(foundedVolumes) > 0 {
+	if len(foundVolumes) > 0 {
 		//We iterate to check if it is exactly the same name
-		for _, foundedVolumeName := range foundedVolumes {
-			if volumeName == foundedVolumeName {
+		for _, foundVolumeName := range foundVolumes {
+			if volumeName == foundVolumeName {
 				return nil, stacktrace.NewError("Volume can not be created because a volume with name '%v' already exists.", volumeName)
 			}
 		}
@@ -61,23 +61,23 @@ func (backend *DockerKurtosisBackend) CreateFileArtifactExpansionVolume(
 	return newFileArtifactExpansionVolume, nil
 }
 
-func (backend *DockerKurtosisBackend) DestroyFileArtifactExpansionVolumes(
+func (backend *DockerKurtosisBackend) DestroyFilesArtifactExpansionVolumes(
 	ctx context.Context,
-	filters *file_artifact_expansion_volume.FileArtifactExpansionVolumeFilters,
+	filters *files_artifact_expansion_volume.FilesArtifactExpansionVolumeFilters,
 ) (
-	map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]bool,
-	map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]error,
+	map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]bool,
+	map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]error,
 	error,
 ) {
-	successfulExpansionVolumeNames := map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]bool{}
-	erroredExpansionVolumeNames  := map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]error{}
+	successfulExpansionVolumeNames := map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]bool{}
+	erroredExpansionVolumeNames  := map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]error{}
 
 	expansionVolumes, err := backend.getMatchingFileArtifactExpansionVolumes(ctx, filters)
 	if err != nil {
 		return nil, nil,  stacktrace.Propagate(err, "An error occurred getting file artifact expansion volumes matching filters '%+v'", filters)
 	}
 
-	//TODO execute in concurrent to improve perf
+	//TODO execute concurrently to improve perf
 	for expansionVolumeName := range expansionVolumes {
 		volumeName := string(expansionVolumeName)
 		if err := backend.dockerManager.RemoveVolume(ctx, volumeName); err != nil {
@@ -100,10 +100,10 @@ func (backend *DockerKurtosisBackend) DestroyFileArtifactExpansionVolumes(
 // ====================================================================================================
 func (backend *DockerKurtosisBackend) getMatchingFileArtifactExpansionVolumes(
 	ctx context.Context,
-	filters *file_artifact_expansion_volume.FileArtifactExpansionVolumeFilters,
-) (map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]*file_artifact_expansion_volume.FileArtifactExpansionVolume, error) {
+	filters *files_artifact_expansion_volume.FilesArtifactExpansionVolumeFilters,
+) (map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]*files_artifact_expansion_volume.FilesArtifactExpansionVolume, error) {
 	if filters == nil {
-		filters = &file_artifact_expansion_volume.FileArtifactExpansionVolumeFilters{}
+		filters = &files_artifact_expansion_volume.FilesArtifactExpansionVolumeFilters{}
 	}
 
 	searchLabels := map[string]string{
@@ -114,14 +114,14 @@ func (backend *DockerKurtosisBackend) getMatchingFileArtifactExpansionVolumes(
 		return nil, stacktrace.Propagate(err, "An error occurred fetching volumes using labels: %+v", searchLabels)
 	}
 
-	matchingObjects := map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]*file_artifact_expansion_volume.FileArtifactExpansionVolume{}
+	matchingObjects := map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]*files_artifact_expansion_volume.FilesArtifactExpansionVolume{}
 	for _, volume := range matchingVolumes {
 		object, err := getFileArtifactExpansionVolumeFromDockerVolumeInfo(
 			volume.Name,
 			volume.Labels,
 		)
 		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred converting volume with name '%v' into a file artifact expansion volume object", volume.Name)
+			return nil, stacktrace.Propagate(err, "An error occurred converting volume with name '%v' into a files artifact expansion volume object", volume.Name)
 		}
 
 		if filters.Names != nil && len(filters.Names) > 0 {
@@ -136,12 +136,6 @@ func (backend *DockerKurtosisBackend) getMatchingFileArtifactExpansionVolumes(
 			}
 		}
 
-		if filters.UserServiceGUIDs != nil && len(filters.UserServiceGUIDs) > 0 {
-			if _, found := filters.UserServiceGUIDs[object.GetServiceGUID()]; !found {
-				continue
-			}
-		}
-
 		matchingObjects[object.GetName()] = object
 	}
 
@@ -151,20 +145,14 @@ func (backend *DockerKurtosisBackend) getMatchingFileArtifactExpansionVolumes(
 func getFileArtifactExpansionVolumeFromDockerVolumeInfo(
 	name string,
 	labels map[string]string,
-) (*file_artifact_expansion_volume.FileArtifactExpansionVolume, error) {
+) (*files_artifact_expansion_volume.FilesArtifactExpansionVolume, error) {
 	enclaveId, found := labels[label_key_consts.EnclaveIDLabelKey.GetString()]
 	if !found {
 		return nil, stacktrace.NewError("Expected the volume's enclave ID to be found under label '%v' but the label wasn't present", label_key_consts.EnclaveIDLabelKey.GetString())
 	}
 
-	guid, found := labels[label_key_consts.GUIDLabelKey.GetString()]
-	if !found {
-		return nil, stacktrace.NewError("Expected to find GUID label key '%v' but none was found", label_key_consts.GUIDLabelKey.GetString())
-	}
-
-	newObject := file_artifact_expansion_volume.NewFileArtifactExpansionVolume(
-		file_artifact_expansion_volume.FileArtifactExpansionVolumeName(name),
-		service.ServiceGUID(guid),
+	newObject := files_artifact_expansion_volume.NewFilesArtifactExpansionVolume(
+		files_artifact_expansion_volume.FilesArtifactExpansionVolumeName(name),
 		enclave.EnclaveID(enclaveId),
 	)