+++ container-engine-lib/lib/backend_interface/objects/file_artifact/filter_artifact_id.go
@@ -1,3 +1,0 @@
-package file_artifact
-
-type FilterArtifactID string