+++ container-engine-lib/lib/backend_interface/objects/files_artifact/files_artifact_id.go
@@ -1,0 +1,3 @@
+package files_artifact
+
+type FilesArtifactID string